+++ src/MappingStrategy/ByTitle.php
@@ -5,6 +5,7 @@ namespace srag\Plugins\Hub2\MappingStrategy;
 use ilObject2;
 use srag\Plugins\Hub2\Exception\HubException;
 use srag\Plugins\Hub2\Object\Category\CategoryDTO;
+use srag\Plugins\Hub2\Object\CompetenceManagement\ICompetenceManagementDTO;
 use srag\Plugins\Hub2\Object\Course\CourseDTO;
 use srag\Plugins\Hub2\Object\CourseMembership\CourseMembershipDTO;
 use srag\Plugins\Hub2\Object\DTO\IDataTransferObject;
@@ -31,12 +32,19 @@ class ByTitle extends AMappingStrategy implements IMappingStrategy {
 			case ($dto instanceof CourseMembershipDTO):
 			case ($dto instanceof GroupMembershipDTO):
 			case ($dto instanceof IOrgUnitMembershipDTO):
+			case ($dto instanceof ICompetenceManagementDTO):
 				throw new HubException("Mapping using Title not supported for this type of DTO");
 				break;
-			case ($dto instanceof GroupDTO):
-			case ($dto instanceof CourseDTO):
-			case ($dto instanceof IOrgUnitDTO):
-			case ($dto instanceof CategoryDTO):
+            case ($dto instanceof IOrgUnitDTO):
+			    $orgu_ids = ilObject2::_getIdsForTitle($dto->getTitle(),'orgu');
+			    foreach($orgu_ids as $orgu_id) {
+			        return $orgu_id;
+                }
+			    break;
+            case ($dto instanceof GroupDTO):
+            case ($dto instanceof CourseDTO):
+            case ($dto instanceof CategoryDTO):
+				//case ($dto instanceof ICompetenceManagementDTO):
 				if ($dto->getParentIdType() != CourseDTO::PARENT_ID_TYPE_REF_ID) {
 					return 0;
 				}
@@ -73,6 +81,8 @@ class ByTitle extends AMappingStrategy implements IMappingStrategy {
 				return "orgu";
 			case ($dto instanceof CategoryDTO):
 				return "cat";
+			/*case ($dto instanceof ICompetenceManagementDTO):
+				return "skmg";*/
 		}
 
 		return '';