+++ src/Sync/Processor/CompetenceManagement/ICompetenceManagementSyncProcessor.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace srag\Plugins\Hub2\Sync\Processor\CompetenceManagement;
+
+use srag\Plugins\Hub2\Sync\Processor\IObjectSyncProcessor;
+
+/**
+ * Interface ICompetenceManagementSyncProcessor
+ *
+ * @package srag\Plugins\Hub2\Sync\Processor\CompetenceManagement
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+interface ICompetenceManagementSyncProcessor extends IObjectSyncProcessor {
+
+}