+++ src/Sync/Processor/OrgUnit/OrgUnitSyncProcessor.php
@@ -148,7 +148,8 @@ class OrgUnitSyncProcessor extends ObjectSyncProcessor implements IOrgUnitSyncPr
 		$parent_id = $this->getParentId($dto);
 
 		$this->current_ilias_object->putInTree($parent_id);
-	}
+        $this->writeRBACLog($this->current_ilias_object->getRefId());
+    }
 
 
 	/**
@@ -190,6 +191,7 @@ class OrgUnitSyncProcessor extends ObjectSyncProcessor implements IOrgUnitSyncPr
 	}
 
 
+
 	/**
 	 * @inheritdoc
 	 *
@@ -204,7 +206,8 @@ class OrgUnitSyncProcessor extends ObjectSyncProcessor implements IOrgUnitSyncPr
 					return;
 				}
 
-				self::dic()->tree()->moveToTrash($this->current_ilias_object->getRefId(), true);
+				$this->current_ilias_object->delete();
+
 				break;
 
 			case IOrgUnitProperties::DELETE_MODE_NONE: